day1/autodatascientist/requirements.txt
kafka-python==1.3.2
numpy==1.12.0
scikit-learn==0.18.1
scipy==0.18.1
